requirements.txt
fastapi==0.132.1
uvicorn==0.41.0
streamlit==1.54.0
requests==2.32.5
langchain==1.2.10
langchain-groq==1.1.2
langchain-mongodb==0.11.0
python-dotenv==1.2.1
pydantic==2.12.5
pymongo==4.16.0